requirements.txt
# Web Framework
fastapi
uvicorn[standard]
pydantic
python-multipart

# Data Processing
pandas
numpy
jinja2

# Date/Time Handling
python-dateutil

# Environment Variables
python-dotenv

# Google AI
google-generativeai

# Google Cloud Services (optional)
google-cloud-pubsub==2.18.4
google-cloud-secret-manager==2.17.0
google-cloud-firestore==2.13.1
google-cloud-aiplatform==1.38.1

# Development and Testing
Faker==20.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

